requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0

Flask-Mail==0.9.1
flasgger==0.9.7.1
PyJWT==2.8.0
cloudinary==1.36.0
Pillow==10.1.0
schedule==1.2.0
